Move Worker into a submodule
